+++ CHANGELOG.md
@@ -9,6 +9,29 @@
 ### Breaking changes
 - Moved the JQuery script tag in `base.html` from block `head_js` to block `vendor_js`
 
+## 20.3.1
+[Full Changelog](https://github.com/uktrade/directory-components/pull/234/files)
+### Bug fixes
+- Re-added MANIFEST.in
+
+## 20.3.0
+[Full Changelog](https://github.com/uktrade/directory-components/pull/227/files)
+
+### Implemented enhancements
+- Added `pagination` template tag
+
+## 20.2.0
+[Full Changelog](https://github.com/uktrade/directory-components/pull/232/files)
+### Implemented enhancements
+- Added new `Search page components` including the `selected filters` and `expandable options` components
+
+## 20.1.0
+[Full Changelog](https://github.com/uktrade/directory-components/pull/226/files)
+
+### Implemented enhancements
+- Added `error_box` and `success_box` template tags
+- Removed MANIFEST.in
+
 ## 20.0.1
 [Full Changelog](https://github.com/uktrade/directory-components/pull/225/files)
 ### Bug fixes